refactor: simplify runner
